feat(server): Separate user_id/username and validate urls
Updated the server endpoint to include user_id in the path and added validation for room_id and user_id. Changed the UpdateRoomData class to UserData and modified the data cleanup logic to use a constant. Updated the .gitignore file to ignore the data directory.


feat(server): Separate user_id/username and validate urls

Updated the server endpoint to include user_id in the path and added validation for room_id and user_id. Changed the UpdateRoomData class to UserData and modified the data cleanup logic to use a constant. Updated the .gitignore file to ignore the data directory.


feat(server): Separate user_id/username and validate urls

Updated the server endpoint to include user_id in the path and added validation for room_id and user_id. Changed the UpdateRoomData class to UserData and modified the data cleanup logic to use a constant. Updated the .gitignore file to ignore the data directory.


feat(server): Separate user_id/username and validate urls

Updated the server endpoint to include user_id in the path and added validation for room_id and user_id. Changed the UpdateRoomData class to UserData and modified the data cleanup logic to use a constant. Updated the .gitignore file to ignore the data directory.